+++ .idea/deploymentTargetDropDown.xml
@@ -14,9 +14,6 @@
       <entry key="ThemeDialogTest">
         <State />
       </entry>
-      <entry key="UpdateShortcutDialogTest">
-        <State />
-      </entry>
       <entry key="app">
         <State />
       </entry>
